+++ lib/routing_client_dart.dart
@@ -1,7 +1,7 @@
 library orsm_dart;
 
-export 'src/osrm_manager.dart';
-export 'src/models/lng_lat.dart';
-export 'src/models/road.dart';
+export 'src/osrm_manager.dart' hide RoadManagerUtils;
+export 'src/models/lng_lat.dart' hide PrvExtLngLat;
+export 'src/models/road.dart' hide PrivateRoad;
 export 'src/utilities/utils.dart'
     hide oSRMServer, GeometriesExtension, OverviewExtension, RoadTypeExtension;