+++ lib/src/models/osrm_mixin.dart
@@ -1,13 +1,12 @@
 import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart' show rootBundle;
-import 'package:routing_client_dart/routing_client_dart.dart';
 import 'package:routing_client_dart/src/models/road_helper.dart';
-import 'package:routing_client_dart/src/utilities/computes_utilities.dart';
 import 'package:routing_client_dart/src/utilities/utils.dart';
-
+/// [OSRMHelper]
+/// 
+/// this helper fpr OSRMManager that contain URL , intruction generator
 mixin OSRMHelper {
-
   String generatePath(
     String server,
     String waypoints, {
@@ -18,7 +17,7 @@ mixin OSRMHelper {
     Geometries geometries = Geometries.polyline,
   }) {
     String url =
-        "$server/routed-${roadType.value}/${profile.name}/v1/diving/$waypoints";
+        "$server/routed-${roadType.name}/${profile.name}/v1/driving/$waypoints";
     var option = "";
     option += "steps=$steps&";
     option += "overview=${overview.value}&";
@@ -50,12 +49,12 @@ mixin OSRMHelper {
     return "$baseGeneratedUrl&source=${source.name}&destination=${destination.name}&roundtrip=$roundTrip";
   }
 
-
   Future<Map<String, dynamic>> loadInstructionHelperJson({
     Languages language = Languages.en,
   }) async {
-    final loadedJson = await rootBundle
-        .loadString('packages/routing_client_dart/src/assets/${language.name}.json', cache: false);
+    final loadedJson = await rootBundle.loadString(
+        'packages/routing_client_dart/src/assets/${language.name}.json',
+        cache: false);
     return json.decode(loadedJson);
   }
 
@@ -131,6 +130,9 @@ mixin OSRMHelper {
       instruction = instructionObject['exit'] as String;
     } else if (name.isNotEmpty && instructionObject.containsKey('name')) {
       instruction = instructionObject['name'] as String;
+    } else if (option['waypointname'] != null &&
+        instructionObject.containsKey('named')) {
+      instruction = instructionObject['named'] as String;
     }
     var firstDestination = "";
     try {